refactor style and start desk auth
