+++ client/root.js
@@ -1,5 +1,5 @@
 import './style/index.scss';
-import Header from './components/header';
+import Header from './components/header/';
 
 export default {
   components: { Header },
