+++ client/views/Auth/signup.vue
@@ -2,21 +2,24 @@
 
 #trusty_auth
 
-  trusty-input(label="new account name")
-    template(slot="input")
-      input
 
-  trusty-input(label="enter email")
-    template(slot="input")
-      input
+  .input_area 
+    .left
 
-  trusty-input(label="enter password")
-    template(slot="input")
-      input
+      trusty-input(label="new account name")
+        template(slot="input")
+          input
 
-  trusty-input(label="confirm password")
-    template(slot="input")
-      input
+      trusty-input(label="enter email")
+        template(slot="input")
+          input
+            
+      confirm-password
+
+    .right
+      ._logo_owl._desk
+        icon-component(name="trusty_owl_small_logo")
+
   
   p._tooltip_p._yellow
     | Write down your password, it CAN'T BE RECOVERED 
@@ -31,17 +34,18 @@
 
   ._bottom_link._margins: span I accept Terms of use
 
-  ._logo_owl
+  ._logo_owl._mob
     icon-component(name="trusty_owl_small_logo")
 
 </template>
 
 <script>
-import trustyInput from '@/components/trusty-input';
+import trustyInput from '@/components/form/input';
 import iconComponent from '@/components/icon';
+import confirmPassword from './confirm-password';
 
 export default {
-  components: { trustyInput, iconComponent }
+  components: { trustyInput, iconComponent, confirmPassword }
 };
 
 </script>
